exellar-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
PyMySQL==1.1.1
PyJWT==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
Werkzeug==3.0.4
gunicorn==22.0.0
cryptography==43.0.1
